refactored spectrogram to work better with NEMS format.
